requirements.txt
flask>=2.3.2
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.3
flask-migrate>=4.0.4
python-dotenv>=1.0.0
httpx>=0.24.1